+++ snakemake/jobs.py
@@ -12,9 +12,9 @@ import json
 import shutil
 import copy
 
-from collections import defaultdict, OrderedDict
-from itertools import chain, filterfalse, groupby
-from operator import attrgetter, itemgetter
+from collections import defaultdict
+from itertools import chain, filterfalse
+from operator import attrgetter
 
 from snakemake.io import (
     IOFile,
@@ -25,6 +25,7 @@ from snakemake.io import (
     get_flag_value,
     wait_for_files,
 )
+from snakemake.resources import GroupResources
 from snakemake.utils import format, listfiles
 from snakemake.exceptions import RuleException, ProtectedOutputException, WorkflowError
 
@@ -1315,7 +1316,18 @@ class GroupJob(AbstractJob):
     @property
     def resources(self):
         if self._resources is None:
-            self._resources = self._calculate_resources()
+            try:
+                self._resources = GroupResources.basic_layered(
+                    toposorted_jobs=self.toposorted,
+                    constraints=self.global_resources,
+                    run_local=self.dag.workflow.run_local,
+                    additive_resources=["runtime"],
+                    sortby=["runtime"],
+                )
+            except WorkflowError as err:
+                raise WorkflowError(
+                    f"Error grouping resources in group '{self.groupid}': {err.args[0]}"
+                )
         return Resources(fromdict=self._resources)
 
     @property
@@ -1514,282 +1526,6 @@ class GroupJob(AbstractJob):
             return self.jobs == other.jobs
         else:
             return False
-
-    def _calculate_resources(self):
-        total_resources = defaultdict(int)
-        total_resources["_nodes"] = 1
-        blocks = []
-        # iterate over siblings that can be executed in parallel
-        for siblings in self.toposorted:
-            # Total resource requirements for this toposort layer
-            block_resources = {}
-
-            job_resources = []
-            pipe_resources = defaultdict(list)
-            for job in siblings:
-                # Get resources, filtering out FileNotFoundErrors. List items will
-                # be job resources objects with (resource: value)
-                # [
-                #   { "runtime": 5, "threads": 2, "tmpdir": "/tmp" },
-                #   { "runtime": 15, "tmpdir": "/tmp"},
-                #   ...
-                # ]
-                # Pipe jobs and regular jobs are put in seperate lists.
-                try:
-                    # Remove any TBDStrings from values. These will typically arise
-                    # here because the default mem_mb and disk_mb are based off of
-                    # input file size, and intermediate files in the group are not yet
-                    # generated. Thus rules consuming such files must thus explicitely
-                    # specify their resources
-                    res = {
-                        k: res
-                        for k, res, in job.resources.items()
-                        if not isinstance(res, TBDString)
-                    }
-                    if job.pipe_group:
-                        pipe_resources[job.pipe_group].append(res)
-                    else:
-                        job_resources.append(res)
-                except FileNotFoundError:
-                    # Skip job if resource evaluation leads to a file not found error.
-                    # This will be caused by an inner job, which needs files created by
-                    # the same group. All we can do is to ignore such jobs for now.
-                    continue
-
-            # Jobs in pipe groups must be run simultaneously, so we merge all the
-            # resources of each pipe group into one big "job". Most resources are
-            # summed, except for cores, which are combined via max().
-            for pipe in pipe_resources.values():
-                job_resources.append(
-                    self._merge_resources(
-                        pipe, default_method=max, methods={"runtime": sum}
-                    )
-                )
-
-            # Set of resource types requested in at least one job
-            resource_types = list(set(chain(*job_resources)))
-            int_resources = OrderedDict()
-            # Sort all integer resources in job_resources into int_resources. Resources
-            # defined as a string are placed immediately into block_resources.
-            for res in resource_types:
-                if res == "_nodes":
-                    continue
-
-                # We use 0 as the default. If we ever want to allow None as an allowable
-                # "2nd" option for string resources (e.g. tmp path is set on some jobs
-                # but not others), this will cause a problem
-                values = [resources.get(res, 0) for resources in job_resources]
-
-                if self._is_string_resource(res, values):
-                    block_resources[res] = values[0]
-                else:
-                    int_resources[res] = values
-
-            # Collect values from global_resources to use as constraints.
-            constraints = OrderedDict(
-                [
-                    (name, self.global_resources.get(name, None))
-                    for name in int_resources
-                ]
-            )
-
-            # For now, we are unable to handle a constraint on runtime, so ignore.
-            # Jobs requesting too much runtime will still get flagged by the
-            # scheduler
-            if "runtime" in constraints:
-                constraints["runtime"] = None
-
-            layers = self._get_layers(int_resources, constraints.values(), "runtime")
-
-            # We extract runtime out of each layer by finding its position in the job
-            # tuple, which will be at the same index as in the OrderedDict
-            # constrained_resources.
-            # Add the max runtime in each layer across all layers
-            if "runtime" in int_resources:
-                runtime_i = list(int_resources).index("runtime")
-
-                runtimes = [max(list(zip(*layer))[runtime_i]) for layer in layers]
-                block_resources["runtime"] = sum(runtimes)
-
-            # Determine the total amount of each resource by summing up each layer,
-            # then taking the max:
-
-            # In each layer, sum across all resource types within the layer,
-            # similar to summing along axis 0 in numpy:
-            # [
-            #   ( 3 ^ , 4 ^ , 1 ),
-            #   ( 2 | , 1 | , 6 ),
-            #   ( 1 | , 4 | , 0 ),
-            # ]
-            sums = [[sum(r) for r in zip(*layer)] for layer in layers]
-
-            # Match up layered resources with definitions in int_resources. Runtime has
-            # already been calculated above via a different method, so we skip it.
-            for i, name in enumerate(int_resources):
-                if name != "runtime":
-                    block_resources[name] = max([layer[i] for layer in sums])
-
-            blocks.append(block_resources)
-
-        if self.dag.workflow.run_local:
-            return Resources(
-                fromdict={**self._merge_resources(blocks, default_method=sum), "_nodes": 1}
-            )
-        else:
-            return Resources(
-                fromdict={
-                    **self._merge_resources(
-                        blocks, default_method=max, methods={"runtime": sum}
-                    ),
-                    "_nodes": 1,
-                }
-            )
-
-    def _is_string_resource(self, name, values):
-        # If any one of the values provided for a resource is not an int, we
-        # can't process it in any way. So we constrain all such resource to be
-        # the same
-        if all([isinstance(val, int) for val in values]):
-            return False
-        else:
-            unique = set(values)
-            if len(unique) > 1:
-                raise WorkflowError(
-                    'Failed to group jobs in group "{group}". Resource {name} '
-                    "is a string but not all group jobs require the same value. "
-                    "Observed values: {values}.".format(
-                        group=self.groupid, name=name, values=unique
-                    )
-                )
-            return True
-
-    def _merge_resources(self, resources, skip=[], methods={}, default_method=max):
-        grouped = {}
-        for job in resources:
-            # Wrap every value in job with a list so that lists can be merged later
-            job_l = {k: [v] for k, v in job.items()}
-
-            # Merge two dicts together, merging key-values found in both into a
-            # list. Code adapted from
-            # https://stackoverflow.com/a/11012181/16980632
-            grouped = {
-                **grouped,
-                **job_l,
-                **{k: grouped[k] + job_l[k] for k in grouped.keys() & job_l},
-            }
-
-        ret = {}
-        for res, values in grouped.items():
-            if res in skip:
-                continue
-
-            if self._is_string_resource(res, values):
-                ret[res] = values[0]
-            elif res in methods:
-                ret[res] = methods[res](values)
-            else:
-                ret[res] = default_method(values)
-        return ret
-
-    def _check_constraint(self, resources, constraints):
-        sums = [sum(res) for res in zip(*resources)]
-        for s, constraint in zip(sums, constraints):
-            if constraint:
-                layers, mod = divmod(s, constraint)
-            else:
-                layers = 1
-                mod = 0
-
-            # If mod not 0, we add 1 to the number of layers. We then subtract
-            # 1, so that if everything fits within the constraint we have 0,
-            # otherwise, some number higher than 0. Finally, we convert to bool.
-            # If the result is 0 or negative, it fits. If greater, it doesn't
-            # fit so we return False
-            if bool(max(0, layers + int(bool(mod)) - 1)):
-                return False
-        return True
-
-    def _get_layers(self, resources, constraints, sortby=None):
-        """Calculate required consecutive job layers.
-
-        Layers are used to keep resource requirements within given
-        constraint. For instance, if the jobs together require 50 threads,
-        but only 12 are available, we will use 5 layers. If multiple constraints are
-        used, all will be considered and met. Any constraints given as None will be
-        treated as infinite.
-        """
-
-        # Calculates the ratio of resource to constraint. E.g, if the resource is 12
-        #  cores, and the constraint is 16, it will return 0.75. This is done for
-        # every resource type in the group, returning the result in a list
-        def _proportion(group):
-            return [r / c if c else 0 for r, c in zip(group, constraints)]
-
-        # Return the highest _proportion item in the list
-        def _highest_proportion(group):
-            return max(_proportion(group))
-
-        rows = [[]]
-
-        # Resources is an OrderedDict: { resource: [val1, val2, val3], ...}.
-        # By zipping, we combine the vals into tuples based on job, 1 tuple per
-        # job: [ (val1, 1_val1, 2_val1), ...]. In each tuple, the resources
-        # will remain in the same order as the OrderedDict, so their identity
-        # can be extracted later.
-        resource_groups = zip(*resources.values())
-
-        # Sort by _proportion highest to lowest
-        pre_sorted = sorted(resource_groups, key=_highest_proportion, reverse=True)
-
-        # If a key is provided (e.g. runtime), we sort again by that key from
-        # highest to lowest
-        if sortby and sortby in resources:
-            # Find the position of the key in the job tuple
-            i = list(resources).index(sortby)
-            pre_sorted = sorted(pre_sorted, key=itemgetter(i), reverse=True)
-
-        for group in pre_sorted:
-            appended = False
-
-            # Check each row for space, starting with the first.
-            for row in rows:
-                if not appended and self._check_constraint(row + [group], constraints):
-                    row.append(group)
-                    appended = True
-
-            # If the final "row" in rows has something, we add a new empty
-            # row. That way, we guarantee we have a row with space
-            if len(rows[-1]) > 0:
-                rows.append([])
-
-            # If not appended, that means a rule required more resource
-            # than allowed by the constraint. This should only be possible for pipe
-            # jobs, which must be run simultaneously.
-            if not appended:
-                too_high = []
-                for i, val in enumerate(_proportion(group)):
-                    if val > 1:
-                        too_high.append(
-                            (list(resources)[i], group[i], list(constraints)[i])
-                        )
-
-                error_text = [
-                    f"\t{res}: {amount}/{constraint}"
-                    for res, amount, constraint in too_high
-                ]
-                raise WorkflowError(
-                    "Not enough resources were provided. This error is typically\n"
-                    "caused by a Pipe group requiring too many resources. Note\n"
-                    "that all resources including 'runtime' are summed across\n"
-                    "every member of the pipe group, except for cores, which is\n"
-                    "calculated via max().\n\n"
-                    "Excess Resources:\n" + "\n".join(error_text)
-                )
-
-        # Remove final empty row. (The above loop ends each cycle by ensuring
-        # there's an empty row)
-        rows.pop()
-        return rows
 
 
 class Reason:
